fixed bug in BufferedPropertyElementTest
